requirements.txt
accelerate==0.28.0
aioboto3==11.3.0
aiobotocore==2.6.0
aiofiles==22.1.0
aiohttp==3.8.6
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.11.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-lru==2.0.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.3
blis==0.7.9
boto3==1.28.17
botocore==1.31.17
bpemb==0.3.5
catalogue==2.0.8
certifi==2023.5.7
certipy==0.1.3
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpathlib==0.16.0
cloudpickle==2.2.1
cmake==3.26.3
comm==0.1.3
confection==0.1.0
conllu==4.5.3
contourpy==1.0.7
cramjam==2.7.0
cryptography==40.0.2
cycler==0.11.0
cymem==2.0.7
dask==2023.6.0
datasets==2.14.5
debugpy==1.6.7
decorator==5.1.1
deeplake==3.8.0
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
emoji==2.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
et-xmlfile==1.1.0
executing==1.2.0
facenet-pytorch==2.5.3
fastjsonschema==2.17.1
fastparquet==2023.8.0
fasttext-wheel==0.9.2
filelock==3.12.0
flair==0.13.1
Flask==3.0.0
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.2.0
funcy==2.0
gdown==5.1.0
gensim==4.3.1
greenlet==2.0.2
huggingface==0.0.1
huggingface-hub==0.22.1
humbug==0.3.2
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itables==1.5.2
itsdangerous==2.1.2
Janome==0.5.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions @ https://github.com/ipython-contrib/jupyter_contrib_nbextensions/tarball/master#sha256=e6a17963a2baded093b95d238f83e2facea84857928739d2d9f04ce2729f4106
jupyter-events==0.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.10.0
jupyter_server_terminals==0.5.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
LeXmo==0.1.4
libdeeplake==0.0.83
lit==16.0.5
littleutils==0.2.2
locket==1.0.0
lxml==5.1.0
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
more-itertools==10.2.0
mpld3==0.5.10
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
NRCLex==3.0.0
numcodecs==0.12.0
numdifftools==0.9.41
numexpr==2.8.4
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openpyxl==3.1.2
outdated==0.2.2
overrides==7.7.0
packaging==23.1
pamela==1.0.0
pandas==2.0.1
pandas-flavor==0.6.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pathos==0.3.1
pathy==0.10.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pingouin==0.5.3
platformdirs==3.5.1
pox==0.3.3
ppft==1.7.6.7
pptree==3.1
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.25.2
psutil==5.9.5
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pybind11==2.11.1
pycparser==2.21
pydantic==1.10.10
Pygments==2.15.1
PyJWT==2.8.0
pyLDAvis==3.4.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-ssim==0.1
pytorch_revgrad==0.2.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
regex==2023.6.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
s3transfer==0.6.2
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
segtok==1.5.11
semopy==2.3.11
semver==3.0.2
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy_fastlang==2.1.0
SQLAlchemy==2.0.15
sqlitedict==2.1.0
srsly==2.4.6
stack-data==0.6.2
stanza==1.7.0
statsmodels==0.14.0
swifter==1.3.5
sympy==1.12
tabula-py==2.9.0
tabulate==0.9.0
terminado==0.17.1
text2emotion==0.0.5
textblob==0.17.1
thinc==8.1.10
threadpoolctl==3.1.0
tika==2.6.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.2.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformer-smaller-training-vocab==0.3.3
transformers==4.39.1
triton==2.2.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.17
vaderSentiment==3.3.2
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.3
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
Werkzeug==3.0.0
widgetsnbextension==4.0.7
Wikipedia-API==0.6.0
wordcloud==1.9.2
wrapt==1.15.0
xarray==2023.10.1
xgboost==2.0.3
xxhash==3.4.1
y-py==0.5.9
yarl==1.9.2
ypy-websocket==0.8.2
zipp==3.15.0
zstandard==0.22.0
